remove bottom section component, rename entity card components
