+++ services/ui-src/src/components/cards/FundingSourcesEntity.tsx
@@ -2,41 +2,22 @@
 import { Heading, Text, Grid, GridItem, Flex } from "@chakra-ui/react";
 import { notAnsweredText } from "../../constants";
 // utils
-import { AnyObject, HeadingLevel, ReportType } from "types";
-import { useStore } from "utils";
+import { AnyObject, HeadingLevel } from "types";
 
-export const EvaluationPlanCard = ({
+export const FundingSourcesEntity = ({
   formattedEntityData,
   headingLevel = "h2",
 }: Props) => {
-  const { report } = useStore() ?? {};
   return (
     <>
       <Heading as={headingLevel} sx={sx.mainHeading}>
-        {formattedEntityData.objectiveName}
+        {formattedEntityData.fundingSource}
       </Heading>
-      <Text sx={sx.subtitle}>
-        Performance measure description or indicators your state or territory
-        will use to monitor progress towards achievement
-      </Text>
-      <Text sx={sx.description}>{formattedEntityData.description}</Text>
-      <Text sx={sx.subtitle}>Performance measure targets</Text>
-      <Text sx={sx.description}>{formattedEntityData.targets}</Text>
-      {report?.reportType === ReportType.WP && (
+      {formattedEntityData.quarters.length > 0 && (
         <>
-          <Text sx={sx.subtitle}>
-            Does the performance measure include quantitative targets?
-          </Text>
-          <Text sx={sx.description}>
-            {formattedEntityData?.includesTargets}
-          </Text>
-        </>
-      )}
-      {formattedEntityData.quarters?.length > 0 && (
-        <>
-          <Text sx={sx.subtitle}>Quantitative Targets</Text>
+          <Text sx={sx.subtitle}>Projected quarterly expenditures</Text>
           <Grid sx={sx.grid}>
-            {formattedEntityData?.quarters.map((quarter: any) => {
+            {formattedEntityData?.quarters?.map((quarter: any) => {
               return (
                 <GridItem key={quarter.id}>
                   <Flex sx={sx.gridItems}>
@@ -57,11 +38,6 @@ export const EvaluationPlanCard = ({
           </Grid>
         </>
       )}
-      <Text sx={sx.subtitle}>
-        Additional detail on strategies/approaches the state or territory will
-        use to achieve targets and/ or meet milestones
-      </Text>
-      <Text sx={sx.description}>{formattedEntityData.additionalDetails}</Text>
     </>
   );
 };
@@ -77,27 +53,12 @@ const sx = {
   mainHeading: {
     fontSize: "md",
   },
-  heading: {
-    fontSize: "sm",
-  },
-  description: {
-    marginTop: "0.25rem",
-    marginBottom: "1rem",
-    fontSize: "sm",
-  },
   grid: {
     display: "grid",
     gridTemplateRows: "1fr 1fr 1fr 1fr",
     gridAutoFlow: "column",
     gridGap: ".5rem",
     marginBottom: "1.25rem",
-  },
-  sarGrid: {
-    display: "grid",
-    gridTemplateRows: "1fr",
-    gridAutoFlow: "column",
-    marginBottom: "1.25rem",
-    width: "50%",
   },
   gridSubtitle: {
     fontWeight: "bold",
@@ -123,10 +84,5 @@ const sx = {
     ".subtitle": {
       marginRight: ".5rem",
     },
-  },
-  unfinishedMessage: {
-    marginY: "1rem",
-    fontSize: "xs",
-    color: "palette.error_dark",
   },
 };
